requirements.txt
altgraph==0.17.4
boto3==1.38.12
botocore==1.38.12
cheroot==10.0.1
defusedxml==0.7.1
fusepy==3.0.1
humanize==4.12.3
jaraco.functools==4.1.0
Jinja2==3.1.6
jmespath==1.0.1
json5==0.12.0
MarkupSafe==3.0.2
more-itertools==10.7.0
packaging==25.0
pefile==2023.2.7
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyperclip3==0.4.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
s3transfer==0.12.0
setuptools==80.8.0
six==1.17.0
tkinterdnd2==0.4.3
urllib3==2.4.0
watchdog==6.0.0
WsgiDAV==4.3.3
